Refactored the internals to correctly render ticks
Improved the "TTL", renamed to "timeInterval" (good for real-time data
that should not be limited to specific timestamps).
Added automatic calculation of the tick frequency based on the given
number of ticks and the time interval).
Refactored the updates and copying the data into the charts for
rendering.
Refactored the data to specify time in milliseconds (integer).
Documentation pending.