Shift the parser buffer to prevent memory leak
Note that the persistent connection's `responseText` buffer should
still be leaking on long-living connections because it is always fully
accessible, but this one is not visible in the Developer Tools.
Note that there is a memory-leak-like behavior visible in Developer
Tools when logging objects to the console (e.g. error objects). It is
not a true memory leak of the application.